refactor(tui): extract submit state helpers
